+++ x-bigint/mod.ts
@@ -1,8 +1,8 @@
 // Copyright 2017-2022 @polkadot/x-bigint authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { xglobal } from 'https://deno.land/x/polkadot@0.0.4-1/x-global/mod.ts';
-
 export { packageInfo } from './packageInfo.ts';
 
-export const BigInt = xglobal.BigInt;
+const _BigInt = BigInt;
+
+export { _BigInt as BigInt };