+++ x-fetch/mod.ts
@@ -1,4 +1,8 @@
 // Copyright 2017-2022 @polkadot/x-fetch authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './browser.ts';
+export { packageInfo } from './packageInfo.ts';
+
+const _fetch = fetch;
+
+export { _fetch as fetch };