+++ x-randomvalues/mod.ts
@@ -1,4 +1,8 @@
 // Copyright 2017-2022 @polkadot/x-randomvalues authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './browser.ts';
+export { packageInfo } from './packageInfo.ts';
+
+export function getRandomValues <T extends Uint8Array> (arr: T): T {
+  return crypto.getRandomValues(arr);
+}