+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## master
+
+- hw-{ledger, ledger-transports}, keyring, networks, util, util-crypto, x-{bigint, fetch, global, randomvalues, textdecoder, textencoder, ws} 9.7.3-5-x
+
 ## 0.0.4-1 Jul 5, 2022
 
 - **beta** Testing changes from latest util & api adjustments for Deno context